generateQRs/requirements.txt
qrcode[pil]
google-cloud-firestore
Pillow
pytz 